+++ _import/tycho-execution/foundry/src/TychoRouter.sol
@@ -511,9 +511,6 @@ contract TychoRouter is AccessControl, Dispatcher, Pausable {
             solverFeeReceiver
         );
 
-        // Finalize all transient deltas to persistent storage
-        _finalizeBalances(msg.sender, tokenIn, amountIn);
-
         amountOut = _maybeAddSolverContribution(
             amountOut, minAmountOut, maxSolverContribution, tokenOut, receiver
         );
@@ -523,6 +520,10 @@ contract TychoRouter is AccessControl, Dispatcher, Pausable {
         } else {
             IERC20(tokenOut).safeTransfer(receiver, amountOut);
         }
+        _updateDeltaAccounting(tokenOut, -int256(amountOut));
+
+        // Finalize all transient deltas to persistent storage
+        _finalizeBalances(msg.sender, tokenIn, amountIn);
 
         _verifyAmountOutWasReceived(
             tokenIn,
@@ -576,9 +577,6 @@ contract TychoRouter is AccessControl, Dispatcher, Pausable {
             solverFeeReceiver
         );
 
-        // Finalize all transient deltas to persistent storage
-        _finalizeBalances(msg.sender, tokenIn, amountIn);
-
         amountOut = _maybeAddSolverContribution(
             amountOut, minAmountOut, maxSolverContribution, tokenOut, receiver
         );
@@ -588,7 +586,10 @@ contract TychoRouter is AccessControl, Dispatcher, Pausable {
         } else {
             IERC20(tokenOut).safeTransfer(receiver, amountOut);
         }
+        _updateDeltaAccounting(tokenOut, -int256(amountOut));
 
+        // Finalize all transient deltas to persistent storage
+        _finalizeBalances(msg.sender, tokenIn, amountIn);
         _verifyAmountOutWasReceived(
             tokenIn,
             tokenOut,
@@ -637,9 +638,6 @@ contract TychoRouter is AccessControl, Dispatcher, Pausable {
             solverFeeReceiver
         );
 
-        // Finalize all transient deltas to persistent storage
-        _finalizeBalances(msg.sender, tokenIn, amountIn);
-
         amountOut = _maybeAddSolverContribution(
             amountOut, minAmountOut, maxSolverContribution, tokenOut, receiver
         );
@@ -649,6 +647,10 @@ contract TychoRouter is AccessControl, Dispatcher, Pausable {
         } else {
             IERC20(tokenOut).safeTransfer(receiver, amountOut);
         }
+        _updateDeltaAccounting(tokenOut, -int256(amountOut));
+
+        // Finalize all transient deltas to persistent storage
+        _finalizeBalances(msg.sender, tokenIn, amountIn);
 
         _verifyAmountOutWasReceived(
             tokenIn,
@@ -953,6 +955,7 @@ contract TychoRouter is AccessControl, Dispatcher, Pausable {
             }
             // Debit the solver's vault balance
             _debitVault(msg.sender, tokenOut, requiredContribution);
+            _updateDeltaAccounting(tokenOut, int256(requiredContribution));
             amount = minAmountOut;
         } else {
             amount = amountOut;